Fix code example outputs + use `print()` function (not statement)
I have fixed the code examples that show incorrect outputs: `Bar -> bar` and `bar* -> bar`.

Also, the `print` statement is now a thing of the past. Code examples should, therefore, rather express the `print()` function usage as to not confuse beginners who in vast majority already use Python 3.x where `print()` is only available as a function now, and not as a statement anymore.